\name{card}
\alias{card}
\docType{data}
\title{
Credit card transactions data for anomaly detection
}
\description{
This data set was taken from Kaggle and contains data of credit card transactions, some of which are frauds.
}
\usage{data("frog")}
\format{
  List with two fields: "normal" and "anomaly" containing respective data frames. Normal contains 10000 rows and anomaly contains 492 rows. There are 29 colums in each data frame.
}
\source{
https://www.kaggle.com/mlg-ulb/creditcardfraud
}
