% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAD-package.R
\docType{package}
\name{GAD-package}
\alias{GAD}
\alias{GAD-package}
\title{GAD: Analysis of Variance from General Principles}
\description{
Analysis of complex ANOVA models with any combination of orthogonal/nested and fixed/random factors, as described by Underwood (1997). There are two restrictions: (i) data must be balanced; (ii) fixed nested factors are not allowed. Homogeneity of variances is checked using Cochran's C test and 'a posteriori' comparisons of means are done using Student-Newman-Keuls (SNK) procedure. For those terms with no denominator in the F-ratio calculation, pooled mean squares and quasi F-ratios are provided. Magnitute of effects are assessed by components of variation.
}
\author{
\strong{Maintainer}: Leonardo Sandrini-Neto \email{leonardosandrini@ufpr.br} (\href{https://orcid.org/0000-0003-4086-6741}{ORCID})

Authors:
\itemize{
  \item Eliandro Gilbert \email{eliandrogilbert@gmail.com}
  \item Maurício Camargo \email{mauricio.camargo@furg.br} (\href{https://orcid.org/0000-0002-8827-1176}{ORCID})
}

}
\keyword{internal}
