\name{normalize}
\alias{normalize}
\title{Normalization function}
\description{
Function for normalizing matrices and data frames.
}
\usage{
normalize(data, max, min)
}
\arguments{
  \item{data}{ data to be normalized. }
  \item{max}{ vector of columns upper bounds. }
  \item{min}{ vector of columns lower bounds. }
}
\value{
  Normalized data.
}
