% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rats.R
\docType{data}
\name{rats}
\alias{rats}
\title{Glycogen content of rat livers}
\format{
A data frame with 36 rows and 4 variables
}
\usage{
data(rats)
}
\description{
Duplicate readings were made on each of three preparations of rat livers
from each of two rats for three different treatments (Sokal & Rohlf, 1995).
}
\details{
\itemize{
\item treat: a fixed factor with three levels
\item rat: a random factor with two levels nested in treat
\item liver: sample size
\item glycog: response variable
}
}
\examples{
library(GAD)
data(rats)
rats
}
\references{
Sokal, R.R., Rohlf, F.J. 1995. \emph{Biometry}: The Principles and Practice of Statistics in Biological Research. 3rd Edition. W.H. Freeman and Co., New York.
}
\keyword{datasets}
